Added banners in ChangeLog to set versions apart.
Entries should be placed in the section for the
version into which they were merged. Entries within
a section should be sorted chronologically.

This is important for keeping track of changes which
weren't merged until a later release than they were
written. (Mixing them in chronologically would cause
confusion about which changes were included in a
release.)