Added empty indra/cmake/CMakeLists.txt to suppress a CMake 2.8 warning.
"CMake Warning (dev) at CMakeLists.txt:36 (add_subdirectory):
The source directory

    /home/jacek/imprudence/linden/indra/cmake

does not contain a CMakeLists.txt file.

CMake does not support this case but it used to work accidentally and
is being allowed for compatibility.

Policy CMP0014 is not set: Input directories must have CMakeLists.txt.
Run "cmake --help-policy CMP0014" for policy details.  Use the
cmake_policy command to set the policy and suppress this warning.
This warning is for project developers.  Use -Wno-dev to suppress it."