RED-565: Building the package/ directory fails for standalone.
See http://redmine.imprudenceviewer.org/issues/565
This patch also reorders the "help" text of llmanifest.py to be
alphabetical, next to adding the option --standalone with help text.
indra/newview/CMakeLists.txt is changed to pass --standalone to
the python script. Finally, viewer_manifest.py, which inherits
from llmanifest.py uses it to skip the sections that don't make
sense when building standalone, and which fail with errors
without this patch.

This patch is not 100% tested: I don't have a Mac or windows.
However, it ONLY touches 'standalone', and before my patches
standalone was broken anyway. So, even if this doesn't fix it
completely for Mac and/or windows, then it can't hurt to commit,
or even release. Even if this fails, the errors are not fatal:
the packaging fails, but the viewer is already compiled at
this point. I'd advise against holding this back until it could
be tested by a Mac or windows developer that can build standalone:
that might never happen to begin with!

When time comes, some other developer will test it and when
there are still problems, fix it, on the other operating systems.