Update of doc/contributions.txt for previous 10 commits.
Oops, I forgot to update doc/contributions.txt in my commits.
This rectifies that. It is handy to have all jira and redmine
labels in doc/contributions.txt so I can see what is already
merged in in a particular source tree (of course, then I'll
have to add those change with the each particulat commit itself,
but I'll assume in this case that all ten will be merged more
or less simultaneously ;).