RED-573: FindTut.cmake is missing: tut.h is not searched for.
This patch only affects --standalone. It was basically ported
from Snowglobe with the following change: If LL_TESTS is not
set, then tut.h is not required (this is an improvement,
because if LL_TESTS is not set no tests should be created or run).
I tested it with -DLL_TESTS:BOOL=ON, -DLL_TESTS:BOOL=OFF and
no -DLL_TESTS* at all, with and without tut installed, internally
and externally. Things worked as expected. Most notably, with
tut.h installed in libraries/include/, it is NOT found. This is
necessary because using -Ilibraries/include for standalone is
a no no: prebuilt stuff might be downloaded, but we don't want
to use it! That directory contains potentially many other libraries.