RED-579: SNOW-751: 'configure' with cmake 2.8 is extremely slow.
Note that indra/cmake/CMakeLists.txt is missing on imprudence, so
I couldn't add 'FindZLIB.cmake' to the list of source files
(see
http://svn.secondlife.com/trac/linden/browser/projects/2009/snowglobe/trunk/indra/cmake/CMakeLists.txt)

I added a comment to
http://github.com/mccabe/imprudence/commit/39f625963e408ec5dd394e202506700259b9b508
that adds an empty indra/cmake/CMakeLists.txt

If Jacek follows up and copies snowglobe's CMakeLists.txt then she'll
have to remove the FindZLIB.cmake from it if she tests it before
merging with THIS commit. In that case it has to be added back when
merging that commit with this one.

Tested on imprudence with cmake 2.8.2:
* Without patch, 'configure' takes 100 seconds.
* With patch, 'configure' takes 4 seconds.