RED-560: Add support for STANDALONE to OPENAL.cmake
Use pkg_check_modules to find freealut (and openal) for STANDALONE.
Non-standalone is untouched by this version.

The changes relative to snowglobe (for the STANDALONE path) is
that I removed the 'include(FindOpenAL) (that comes with cmake)
which not only doesn't find libalut, but even sets the wrong
cmake variables (it sets OPENAL_LIBRARY, and our cmake scripts
*use* OPENAL_LIBRARIES). Moreover, under the assumption that
STANDALONE really is only for linux (and I think that is so),
every linux distribution DOES have freealut (and openal) installed
with a pkg-config anyway. I also removed pkg_check_modules(openal)
because it is redundant: freealut pulls that in already.