RED-429: Additional fixes.
This patch fixes the regression that you couldn't run
'scripts/install.py' anymore from linden/ when installing a prebuilt
package.

It also adds support for --dry-run: still printing what symlinks are
missing when that options is used without actually adding the symlinks.

Finally, the added symlinks are now added to the 'files' section in
install.xml so that they are also removed again when you uninstall the
package.