IMP-661: WARNING: isFeatureAvailable: Feature RenderCubeMap not on feature list!
Fixes this warning. Note that LLCubeMap::sUseCubeMaps is set to true by
default, so this patch only has effect when the feature is actually NOT
available.

I tested that LLCubeMap::sUseCubeMaps is NOT used before the point where
it is initialized now.

The patch also adds an entry for IMP-660 to doc/contributions.txt that I
had forgotten in the previous commit; and while I was at it, changes all
RED- prefixes in IMP-.