Add headers from libdbus-1-dev_1.2.1-5+lenny1_i386.deb to dbusglib linux 32 bit package.
The dbusglib package contains headers from both, libdbus-1 as well as
libdbus-glib-1, but only the shared object file of the latter.
This isn't entirely consistent with how we do this for other packages
that we want to use the system provided ones for: then we link against
prebuilts, but just don't install them. However, if that is working
then it's very likely that this method will work too for a C library,
so I guess it's ok like this.

This package now contains the shared library from
libdbus-glib-1-2_0.76-1_i386.deb (plus symlinks) and the headers from
libdbus-glib-1-dev_0.76-1_i386.deb AND libdbus-1-dev_1.2.1-5+lenny1_i386.deb.