Rename hippo* files to lower case (PART 1 OF 2)
There was a problem with #include "hippolimits.h" because
the file is called hippoLimits.h (with an uppercase L).
After some discussion on IRC it was reluctantly decided that
filenames should be lowercase, as is done for every Linden file.
Moreover, hippolimits.h is about the class HippoLimits, with
uppercase H, so even that doesn't match.

Of course, then we immediately rename all hippo* files to
lowercase.

This patch ONLY renames the file, with no other changes.
This is done in order to make sure that git won't get confused ;).
All other necessary changes will be in the next commit.