Rename hippo* files to lower case (PART 2 OF 2)
There was a problem with #include "hippolimits.h" because
the file is called hippoLimits.h (with an uppercase L).
After some discussion on IRC it was reluctantly decided that
filenames should be lowercase, as is done for every Linden file.

This is part 2 or 2 (the first part just renamed the files).
This part changes all files to make the viewer compile again
after the renaming.