IMP-592: Extend login screen menu bar till the top of the window.
Basically, this Linden code is just horribly broken.
I spend already too much time on this (two days) for dead code, so I
settled for a minimal fix for this particular case that works, instead
of fixing their code properly.

The problem is that without specifying "bottom" the code adds 4 pixels
(VPAD) between whatever widget that is added and it's parents top
(in LLView::createRect). In order to be sure that bottom is exactly
minus the height, I had to specify the height too; hopefully this
value is the same for everyone cause theorectically it depends on the
font size (4 + fontheight + 4). The default height of the Linden
code is only 4 + fontheight (they leave off the VPAD).