Partial fix #772: Large sculpt maps are deformed
This is a partial fix for cases where the sculpt map is larger than
128x128, like in some old sculpted trees. This was done by content
creators to beat bugs in old viewers. The sculpties still need a lod
change to trigger a rebuild once they are fully decoded, and that doesn't
always happen, because we abuse RenderVolumeLODFactor.

Ported from Snowglobe2 (GPL)