Ported viewerinfo.cpp from Kokua.
It defines the ViewerInfo module/namespace, which contains a variety
of functions for querying the viewer name and version. It is designed
to be a replacement for viewerversion.cpp and viewerversion.xml.

viewerversion.cpp has been removed, but viewerversion.xml remains
(albeit with a bogus version number) until the packaging system no
longer uses it.