disable voice if ParcelVoiceInfoRequest response is malformed.
Seen (empty LLSD) in OSGrid Mumble Sandbox. Without a "voice_credentials" key the client never connects to the server, but would retry in vain to do so.
Note to OpenSim devs: the "channel_uri" of the ParcelVoiceInfoRequest response is parsed to identify if we are connecting to a mumble server
on parcel change. The presence of "sip:" within the uri is interpreted as connecting to vivox/freeswitch, the absence as connecting to a mumble server. However it would be much better to add a "server_type" key to the
response, which could hold a string like "mumble", "freeswitch", etc -
the absence would indicate connecting to a legacy(like) voice server.
Additional note: to detect a mumble server also of the  VoiceAccountProvision response the voice_account_server_uri is parsed:
If the uri starts with tcp:// the "mumble" client is loaded(by default, you can change it in the "VoiceModuleMumble" debug setting), for any other scheme "SLVoice" (default, "VoiceModuleVivox" debug setting).